changed type in store
